move StackableContentPanel component to folder
