+++ src/modules/stackables/components/stackableContentPanel/StackableContentPanel.js
@@ -4,7 +4,7 @@
 import { html, nothing } from 'lit-html';
 import { repeat } from 'lit-html/directives/repeat.js';
 import css from './stackableContentPanel.css';
-import { MvuElement } from '../../MvuElement';
+import { MvuElement } from '../../../MvuElement';
 
 const Notification_Autoclose_Time = 5000;
 const Update_Notifications = 'update_notifications';
