+++ src/modules/stackables/components/stackableContentPanel/stackableContentPanel.css
